pyproject.toml
[project]
name = "langchain_dartmouth"
dynamic = ["version"]
authors = [
  { name="Simon Stone", email="simon.stone@dartmouth.edu" },
]
description = "LangChain components for Dartmouth-hosted models."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "dartmouth-auth",
    "langchain",
    "langchain_community",
    "langchain_huggingface",
    "huggingface-hub==0.26.3",
    "langchain_openai",
    "text_generation",
    "python-dotenv",

]
license = {file = "LICENSE"}

[project.optional-dependencies]
docs = ["sphinx", "sphinx-rtd-theme", "autodoc_pydantic"]
dev = [
  "pytest",
]

[project.urls]
Homepage = "https://github.com/Dartmouth-Libraries/langchain-dartmouth"
Issues = "https://github.com/Dartmouth-Libraries/langchain-dartmouth/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths =[
   "tests"
]

[tool.hatch.version]
path = "src/langchain_dartmouth/_version.py"
